Moved EventSource to Framework
Makes it accessible to more files.
